requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
crewai==1.5.0
cryptography==46.0.3
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docxtpl==0.20.2
dotenv==0.9.9
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.5
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
json5==0.12.1
json_repair==0.25.2
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.22.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
nodeenv==1.9.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.0
portalocker==2.7.0
postgrest==2.24.0
posthog==5.4.0
pre_commit==4.4.0
propcache==0.4.1
protobuf==6.33.1
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
realtime==2.24.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.29.0
rsa==4.9.1
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
tomlkit==0.13.3
tqdm==4.67.1
ty==0.0.1a27
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uv==0.9.11
uvicorn==0.38.0
virtualenv==20.35.4
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
